better handle cache on static iframes
